Rearrange methods in Table and Base. Add more test cases for testing Base
