+++ tests/schemata/schema1/__init__.py
@@ -2,6 +2,4 @@ __author__ = 'eywalker'
 import datajoint as dj
 
 print(__name__)
-from .test1 import *
-from .test2 import *
 from .test3 import *
