+++ tests/schemata/schema2/__init__.py
@@ -1,1 +1,2 @@
 __author__ = 'eywalker'
+from .test1 import *
