+++ tests/schemata/schema2/test1.py
@@ -8,9 +8,8 @@ import datajoint as dj
 
 
 class Subjects(dj.Base):
-    _table_def = """
-    test2.Subjects (manual)     # Basic subject info
-
+    definition = """
+    schema2.Subjects (manual)     # Basic subject info
     pop_id       : int      # unique experiment id
     ---
     real_id                     :  varchar(40)    #  real-world name
