+++ tests/test_utils.py
@@ -3,7 +3,6 @@ Collection of test cases to test core module.
 """
 
 __author__ = 'eywalker'
-from . import (CONN_INFO, PREFIX, BASE_CONN, cleanup)
 from nose.tools import assert_true, assert_raises, assert_equal
 from datajoint.utils import to_camel_case, from_camel_case
 from datajoint import DataJointError
